Before:
```
fz@centrifuge: go test -test.run=XXX -bench=BenchmarkBroadcast -cpuprofile=cpu -memprofile=mem
goos: darwin
goarch: amd64
pkg: github.com/centrifugal/centrifuge
BenchmarkBroadcastMemoryEngine/JSON_1_subscribers-4         	  200000	      7901 ns/op	     944 B/op	      26 allocs/op
BenchmarkBroadcastMemoryEngine/Protobuf_1_subscribers-4     	  300000	      4870 ns/op	     465 B/op	      14 allocs/op
BenchmarkBroadcastMemoryEngine/JSON_10000_subscribers-4     	     100	  11249143 ns/op	  481196 B/op	   20025 allocs/op
BenchmarkBroadcastMemoryEngine/Protobuf_10000_subscribers-4 	     100	  12077671 ns/op	  480677 B/op	   20013 allocs/op
```
After:
```
fz@centrifuge: go test -test.run=XXX -bench=BenchmarkBroadcast -cpuprofile=cpu -memprofile=mem
goos: darwin
goarch: amd64
pkg: github.com/centrifugal/centrifuge
BenchmarkBroadcastMemoryEngine/JSON_1_subscribers-4         	  200000	      7698 ns/op	     896 B/op	      24 allocs/op
BenchmarkBroadcastMemoryEngine/Protobuf_1_subscribers-4     	  300000	      4415 ns/op	     417 B/op	      12 allocs/op
BenchmarkBroadcastMemoryEngine/JSON_10000_subscribers-4     	     100	  11217320 ns/op	    1097 B/op	      24 allocs/op
BenchmarkBroadcastMemoryEngine/Protobuf_10000_subscribers-4 	     100	  18144938 ns/op	     828 B/op	      12 allocs/op
```